Labels and extension metadata
This thread came from an issue that is still open [1], so I'm tagging
it.  My personal feeling is still "this is out-of-scope, use an
orthogonal spec and load it in as arbitrary bundle content [2]" [3].
But the issue is still alive and fluxy, and I was just talking about
it with Vincent in #opencontainers on 2015-12-18.

[1]: https://github.com/opencontainers/specs/issues/108
[2]: Message-ID: <20150826195447.GX21585@odin.tremily.us>
     Subject: Dropping the rootfs requirement and restoring arbitrary bundle content
     Date: Wed, 26 Aug 2015 12:54:47 -0700