runtime-spec: Linux: Definining "container" and "container processes"
With [1], which landed on 2016-05-26 [2], 'kill' is signalling "the
process in the container".  And since [3] (2017-06-04), the only
remaining "processes" references have to do with namespaces, where the
namespace in question is fairly clear.  So I don't think there's much
uncertainty left about what "container processes" means, or any
implication that it's a single set of processes.

[1]: https://github.com/duglin/runtime-spec/blob/be594153b522f52bebd0500cef6fe0f1d77f6a59/runtime.md#kill
[2]: https://github.com/opencontainers/runtime-spec/pull/384#event-673882236
[3]: https://github.com/opencontainers/runtime-spec/pull/809#event-1123388834